requirements.txt
asgiref==3.8.1
certifi==2024.6.2
Django==5.0.2
django-elasticsearch-dsl==8.0
djangorestframework==3.15.1
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.13.1
mysqlclient==2.2.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
